requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-property==0.2.2
asyncstdlib==3.12.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl-cffi==0.5.10
duckduckgo_search==4.2
fastapi==0.109.0
Flask==3.0.0
frozenlist==1.4.1
g4f==0.2.0.4
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
loguru==0.7.2
lxml==5.1.0
lz4==4.3.3
MarkupSafe==2.1.3
mmh3==4.1.0
multidict==6.0.4
nest-asyncio==1.5.9
outcome==1.3.0.post0
pillow==10.2.0
platformdirs==4.1.0
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
PyExecJS==1.5.1
PySocks==1.7.1
python-telegram-bot==20.7
requests==2.31.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.35.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
undetected-chromedriver==3.5.4
urllib3==2.1.0
uvicorn==0.26.0
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.4